Move SceneNode definition to it's own .h and .cpp
