+++ src/smol_transform.cpp
@@ -1,5 +1,5 @@
 #include <smol/smol_transform.h>
-#include <smol/smol_scene.h>
+#include <smol/smol_scene_nodes.h>
 #include <smol/smol_systems_root.h>
 
 namespace smol
